drm/panic: Move drawing functions to drm_draw
[ Upstream commit 31fa2c1ca0b239f64eaf682f1685bbbd74fc0181 ]

Move the color conversions, blit and fill functions to drm_draw.c,
so that they can be re-used by drm_log.
drm_draw is internal to the drm subsystem, and shouldn't be used by
gpu drivers.

Link: https://patchwork.freedesktop.org/patch/msgid/20241204160014.1171469-2-jfalempe@redhat.com
Stable-dep-of: 05663d88fd0b ("drm/tests: Fix drm_test_fb_xrgb8888_to_xrgb2101010() on big-endian")
(cherry picked from commit 268a5b682a18ac8b79b90c3abec3707db408464f)
